Major janitoring...
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@10516 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/GUIDialogFileBrowser.cpp
@@ -197,7 +197,7 @@ bool CGUIDialogFileBrowser::OnMessage(CGUIMessage& message)
     break;
   case GUI_MSG_SETFOCUS:
     {
-      if (m_viewControl.HasControl(message.GetControlId()) && m_viewControl.GetCurrentControl() != message.GetControlId())
+      if (m_viewControl.HasControl(message.GetControlId()) && (DWORD) m_viewControl.GetCurrentControl() != message.GetControlId())
       {
         m_viewControl.SetFocused();
         return true;
